Moved test from main.rs to types.rs
